// apps/admin/src/api/generated/schemas/activityLogStats.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * x-ear API
 * Backend API for x-ear CRM
 * OpenAPI spec version: 1.0.0
 */
import type { ActivityLogStatsTopActionsItem } from './activityLogStatsTopActionsItem';

/**
 * Activity log statistics
 */
export interface ActivityLogStats {
  /** Total logs */
  total?: number;
  /** Logs in last 24 hours */
  last24Hours?: number;
  /** Critical logs count */
  critical?: number;
  /** Top actions */
  topActions?: ActivityLogStatsTopActionsItem[];
}